Refactored Table class's Set/Get methods, added RawGet/Remove methods
Refactored the Table class's Set/Get methods to be more symmetrical.
There is a method overload for string, integer, DynValue and object keys
as well as a method with multiple keys for easier lookup in sub-tables.

Added a set of RawGet methods that return null if item is not found.

Added a set of Remove methods to remove a key from the Table object.

Added an Append method to append a value to the table using the next
available integer index.